+++ ploneapi_shell/__init__.py
@@ -1,5 +1,5 @@
 """ploneapi-shell package."""
 
 __all__ = ["__version__"]
-__version__ = "0.1.17"
+__version__ = "0.1.18"